% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastliu-package.R
\docType{package}
\name{fastliu-package}
\alias{fastliu}
\alias{fastliu-package}
\title{fastliu: Fast Functions for Liu Regression with Regularization Parameter and Statistics}
\description{
Efficient computation of the Liu regression coefficient paths, Liu-related statistics and information criteria for a grid of the regularization parameter. The computations are based on the 'C++' library 'Armadillo' through the 'R' package 'Rcpp'.
}
\author{
\strong{Maintainer}: Murat Gen<U+00E7> \email{muratgenc@tarsus.edu.tr} (\href{https://orcid.org/0000-0002-6335-3044}{ORCID})

Authors:
\itemize{
  \item <U+00D6>mer <U+00D6>zbilen \email{ozbilen@mersin.edu.tr}
}

}
\keyword{internal}
