% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoliu.R
\name{infoliu}
\alias{infoliu}
\title{Information Criteria for Liu Regression}
\usage{
infoliu(obj)
}
\arguments{
\item{obj}{A \code{liureg} object}
}
\value{
\code{infoliu} returns the matrix of information criteria
for each value of the regularization parameter \code{lambda}.
}
\description{
For each value of \code{lambda}, infoliu calculates
the values of the AIC and BIC model selection criteria.
Model selection criteria are based on the degrees of the freedom,
\eqn{\texttt{df}=\mathrm{trace}\left(\mathbf{H}_\lambda\right)}
of the Liu regression model where \eqn{\mathbf{H}} is the hat matrix of
Liu regression model.
}
\examples{
data("Hitters")
Hitters <- na.omit(Hitters)
X <- model.matrix(Salary ~ ., Hitters)[, -1]
y <- Hitters$Salary
lam <- seq(0, 1, 0.01)
liu.mod <- liureg(X, y, lam)
infoliu(liu.mod)
}
\references{
Akaike, H. (1974). A new look at the statistical model identification.
\emph{IEEE Transaction on Automatic Control}, \strong{9}(6), 716-723.
\doi{10.1109/TAC.1974.1100705}.

Liu, K. (1993). A new class of blased estimate in linear regression.
\emph{Communications in Statistics-Theory and Methods}, \strong{22}(2), 393-402.
\doi{10.1080/03610929308831027}.

Schwarz, G. (1978). Estimating the dimension of a model.
\emph{Annals of Statistics}, \strong{6}(2), 461-464.
\doi{10.1214/aos/1176344136}.
}
\seealso{
\code{\link[=predict]{predict()}}, \code{\link[=summary]{summary()}}
}
\author{
Murat Genç
}
